requirements.txt
altgraph==0.17.4
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
et_xmlfile==2.0.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
num2words==0.5.14
openpyxl==3.1.5
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.9.0
PyMuPDF==1.26.0
pytesseract==0.3.13
pywin32-ctypes==0.2.3
requests==2.32.3
scapy==2.6.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
